+++ web/src/actions/editActivity/index.js
@@ -1,3 +1,4 @@
+export * from './activityList';
 export * from './contractorResources';
 export * from './costAllocate';
 export * from './goals';
